chore(*): do not test for `.ts` & `.tsx`
this is rather sad but i don't have a solution :(

reason is that `.ts` & `.tsx` files are parsed with
`react-docgen-typescript` which uses `typescript`

`typescript.createProgram(path)` is used where `path` from mocked file
system does not correspond to real file. as a result, typescript doesn't
have anything to parse and so my expectations are not fulfilled.

in real life it works, because file system is not mocked. But we don't
have test coverage.

one possible approach could be to mock file system not only in memory
(as currently) but use actual filesystem under `/tmp` or similar.