chore(followProps): ignore `dist/` in composed paths
consider this case:

```js
import {Text as CoreText} from 'wix-ui-backoffice/dist/src/components/core/CoreText';

// then
static propTypes = {
  ...CoreText.propTypes,
}
```

when this happens, react-docgen tries to read transpiled code which is
not something that it can understand. Thus we're making an assumption
that there's also `src` in published package and we can read source from
there.